feeders/r-sim-feeder/requirements.txt
# FIXME: This is just a manual list currently of major requirements.
#        Needs later conversion to proper packaging list.
asyncio
websockets
json
apache-iotdb